Add several fixes

+++ objects/shell-scripts/install.sh
@@ -3,5 +3,5 @@
 CONF_DIR=$(realpath `dirname $0`)
 
 mkdir -p $HOME/bin/
-ln -sf $CONF_DIR/scripts $HOME/bin/xdf
+[ ! -L "$HOME/bin/xdf" ] && ln -sf $CONF_DIR/scripts $HOME/bin/xdf
